Split `main` function code into separate functions
